# server/app/model/role.go
package model

import "time"

// Role merepresentasikan peran pengguna dalam aplikasi
type Role struct {
	ID        uint      `gorm:"primaryKey" json:"id"`
	Title     string    `json:"title"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}